chore(apps/interface): refactor TokenCard
